backend/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.100.0
greenlet==3.0.1
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
passlib==1.7.4
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.8
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.22.0
